requirements.txt
influxdb==5.3.0
python-dateutil==2.8.1
speedtest-cli==2.1.2
